Develop

+++ src/main/java/com/example/bot/_for_shelter/repository/RoomRepository.java
@@ -7,9 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     boolean existsByIdForEntry(int randomNumber);
 
     @Query("SELECT r FROM Room r WHERE r.idForEntry = :idForEntry")
-    Optional<Room> findByIdForEntry(int idForEntry);
+    Optional<Room> findByIdForEntry(long idForEntry);
+
 }
